refactor: make the code more readable
